ENH Add send, throw, and close to JsProxy of generators (#3294)
This adds two new types to `pyodide.ffi`: `JsIterator` (which requires an object
to have a `next` method) and `JsGenerator` (which requires an object to be a true
JavaScript generator). The `JsIterator` type has `__next__` and `send` implementations,
and `JsGenerator` has also `throw` and `close` methods.

Unlike for Python generators, it's allowed to `send` a non-None value into a not-started
generator. We have no way to detect that the generator isn't started (JavaScript just 
ignores the value).

This also improves the error messages when `next` called on an object returns something
unexpected.