Add test for MutableMapping methods on object_maps and fix bug (#3297)
This adds tests of all the `MutableMapping` methods for objects returned by 
`as_object_map`. They all worked but I was accidentally adding only the non 
mutating methods. I also added `JsMap` and `JsMutableMap` to `pyodide.ffi` 
and fixed it so that `isinstance` works correctly and mypy agrees with the usage.