Add athrow and aclose to JsProxy of an AsyncGenerator (#3299)
Continuing with my work on support for generators, this adds `athrow` and 
`aclose` for async generators.

Currently if you use `athrow` and the generator doesn't catch the error, it will
throw a double wrapped error. This is probably also a problem with async calls
Python --> Js --> Python because the way that we control the Python exception
lifetime only works correctly in synchronous contexts. Fixing this is a TODO.

Also, there is still a question of what to do with return values of async Javascript
generators. Currently I think they are available from Python as 
`async_stop_iteration_exception.args[0]`.