Add generator argument lifetime handling (#3317)
This adds lifetime handling to the arguments of JavaScript sync and async generator
functions and to the arguments of `send` and `asend`. The arguments and the `send`
values are kept alive until the generator finishes. This is accomplished through a
wrapper around the generator. The return value of the generator is not destroyed by
the wrapper, but it will get destroyed in the C code for `asend`/`athrow`/`aclose` (unless
the generator wrapper is sent back to JavaScript and called directly).