DOCS: Help Spinx ModuleAnalyzer find ffi types and fix for overloaded functions
Two fixes:
1.
the `pyodide.ffi` stuff is defined in `_pyodide._core_docs`.
We don't want `_pyodide._core_docs` to appear in the documentation
because this isn't where you should import things from so we override
the `__name__` of `_pyodide._core_docs` to be `pyodide.ffi`. But then
Sphinx fails to locate the source for the stuff defined in
`_pyodide._core_docs`. This patches `ModuleAnalyzer` to tell it to
look for the source of things from `pyodide.ffi` in
`_pyodide._core_docs`.

2.
A second problem is that if a function has a typing.overload then
the code path that sphinx-autodoc takes is not one that
sphinx-autodoc-typehints can handle.
See https://github.com/tox-dev/sphinx-autodoc-typehints/issues/296