Add fixes for attribute documentation
autodoc emits an event autodoc-process-docstring when it has processed
the docstring of a function or method which sphinx-autodoc-typehints
uses to reformat the types. However, no event is emitted when an
attribute is processed and the attribute type annotations are not
consistent with the method type annotations.

To fix this, we monkeypatch stringify_typehint. But then there is a
second problem which is that the `:type:` directive for `PyAttribute`
does not process rst, so we monkeypatch that too.