Rework PyProxy
There are three stages in which we are concerned with the behavior of the objects we define:
1. at time of static typechecks
2. at execution time
3. when generating docs

Prior to this PR, the subtypes of PyProxy only work well for static type checks, they work
acceptably at runtime (just don't let the user access them), and the docs don't look that
great. This PR is primarily intended to improve the docs for PyProxy, but they also make
execution time checks work better: you can now say `obj instanceof pyodide.PyCallable`
instead of `obj.isCallable()` which I is easier to understand and to cross reference
against the documentation. I am marking `isCallable` as deprecated.

I also made a bunch of edits and improvements to the docs.

I have deprecated `PyProxyCallable` in favor of `PyCallable` and `isCallable` in favor of
`obj instanceof PyCallable`.

One annoyance here is the presence of `PyBuffer`. I would like to rename the current `PyBuffer`
to something else (`PyBufferView`?) and rename the thing I have currently called `PyProxyBuffer`
to `PyBuffer`. This may need to happen over a few releases to make time for deprecation warnings.

Another thought: perhaps we could rename `JsProxy` to `JsObject` and `PyProxy` to `PyObject`?