Add an argument to loadPyodide to pass environment variables
This also updates the command line runner to pass in all ambient
environment variables.

I added handling for `homedir` and `env.HOME`: if both are passed,
raise an error, otherwise they mean the same thing.

This changes the behavior of HOME in the command line runner: it
used to point to the current working directory, but now it will
point to the ambient home directory.