Replace void return value with int in quadpack.h (#3953)
In C++
```
try {
	some_func(arg1, arg2, ...);
} catch (SomeException e) {
	// ...
}
```
generates code like:
```
invoke_sig(some_func_ptr, arg1, arg2, ...);
```
where invoke_sig is a JS trampoline that does a JS try/catch. So in particular, if some_func is declared with the wrong type but it is only called in try blocks, then it will work anyways. But if instead of js exceptions you use wasm exceptions then it's back to "null function or function signature mismatch".