ENH Performance Add pyproxy_lookup for $$ instead of accessing it through proxy traps
We've received user feedback recently that Python => JS calls are a bit slow when the
arguments are proxied. Profiling this showed that 90% of the time was being spent in
destroying the pyproxies. The line profiles showed that every time we accessed
`pyproxy.$$` it was very slow. For normal JS objects, attribute access is very efficient
but for Proxy exotic objects with `get` traps it's pretty slow. The cool kids use `WeakMap`
to avoid this problem.

This patch was enough to bring the time spent in destroy_proxy down from 90% to 6% of the
runtime for the following profile:
```py
from pyodide.ffi import create_proxy
from pyodide.code import run_js
from timeit import timeit
n = 100000

jsf = run_js("() => {}")
a = object()

timeit("jsf(a, a, a, a)", number=n, globals={"jsf": jsf, "a": a}) / n * 1e6
```

To keep this patch small, I left in the old ways of accessing the `$$` object via `pyproxy.$$`.
In a followup we should switch everything else over to using the map, and also move
`$$flags` and `$$props` into the map.