PERF Only gc register proxies if not destroyed at end of function call (#4025)
`finalizationRegistry.register` is expensive. This PR avoids calling it when converting function arguments. If the function returns a promise or generator and the lifetime of the arguments is extended, then we register all the arguments in the `finalizationRegistry`. Otherwise, we immediately destroy the arguments so they don't need to be registered.