ENH Add `response.text` to `pyodide.http.FetchResponse`
We've received feedback from users that use other requests APIs that
they expect the method to be called `response.text` instead of
`response.string`. Indeed both the Fetch response API and the
Python requests library use this convention:

https://developer.mozilla.org/en-US/docs/Web/API/Response/text
https://requests.readthedocs.io/en/latest/api/#requests.Response.text

This adds `response.text` to `FetchResponse`. It is a synonym for
`response.string`.