PERF Improve internal PyProxy attribute access (#4014)
We've received user feedback recently that Python => JS calls are a bit slow when
the arguments are proxied (from @DerThorsten and @laffra). Profiling showed
that 90% of the time was being spent in destroying the pyproxies. The line profiles
showed that every time we accessed `pyproxy.$$` it was very slow. The problem is
that these accesses go through a slow path in the `PyProxy` traps for `get`. Switching
to using a `Symbol` for access hits a fast path for the `get` trap.

This patch was enough to bring the time spent in destroy_proxy down from 90% to 6%
of the runtime for the following profile:
```py
from pyodide.ffi import create_proxy
from pyodide.code import run_js
from timeit import timeit
n = 100000

jsf = run_js("() => {}")
a = object()

timeit("jsf(a, a, a, a)", number=n, globals={"jsf": jsf, "a": a}) / n * 1e6
```