PERF Use a slotmap for hiwire
A slot map is a map that generates and returns the key when inserting a
value. Slot maps have faster lookup, insertion, and deletion than hash maps,
though they aren't built in to the language so they are a bit more effort to
implement. The implementation here is based roughly on this Rust file:
https://github.com/orlp/slotmap/blob/master/src/basic.rs