Improve handling of generated typescript files (#4212)
This makes IDEs, documentation generation tools, and linters handle our
generated files better. I set the project root to be `src` instead of `src/js`  
so that it is allowed to import files directly from `src/core`. This way we 
don't have to copy `error_handling.ts` we can just import
`../core/error_handling`.

I made a new folder called `src/js/generated` to place generated files into and
added typescript resolution rules so that when we import a file called
"generated/blah" we first look for `blah` in `src/js/generated` and then fall
back to a file called `blah` in `src/core`.

This also allows us to move around fewer files when building the docs 
and in the makefile.