Fix static_import test to work in safari (#4296)
The test_static_import test was failing in Safari because it runs another Safari session
while Safari does not allow multiple sessions running simultaneously.

We already have a workaround for it, but it was not working for this test, because it
was not using selenium fixture directly. This PR refactors the test, so our workaround can be applied.