Split build-recipes and build-recipes-no-deps into separate commands
There are a lot of options that are only implemented for one or the other of
these. We warn about some combinations of options that are ignored but skip a
lot of them. Seems best to turn them into separate commands so that the two
commands each only accept arguments that change their behavior