Reduce the amount of stuff that the makefile always redoes
libgl is marked .PHONY and is a dependency for the link step which forces us to
always redo the link step even though it's completely not necessary. I moved
`embuilder build libgl` into the link step since that's the only place where
it's needed.

`dist/python_stdlib.zip` and `dist/pyodide-lock.json` both declare a dependency
on `pyodide_build` but there is never a file of that name so they always have to
run. Instead, run `make pyodide_build` so that they can reuse the logic but they
don't always run if their other deps are satisfied.

Instead `dist/python_stdlib.zip` needs to rely on all files in `src/py`.

Make `pyodide_build` only invoke `pip install` if `pyodide_build` is not already
installed. Pip generates a lot of noise and takes a while so skipping it when we
don't need it helps a lot.