Add error handling to mountNativeFS (#4559)
Most of these cases it would have thrown an error all along and this just adds a clearer error message. The only case where this adds a new error is if the directory is nonempty but is not a file system root. But in that case, it's unexpected that it overwrites the original directory contents. The documentation says:

> If it does exist, it must be empty.

So people might reasonably expect us to throw an error in that case rather than silently overwriting.