FIX don't let promising call's exception be stolen by interleaved Python calls
Consider the code:
```js
pyodide.runPython(`
    def a():
        raise Exception("hi")
    def b():
        return 7;
`);
const a = pyodide.globals.get("a");
const b = pyodide.globals.get("b");
const p = a.callSyncifying();
assert(b() === 7);
await p;
```

This used to misbehave because because a()'s error status got stolen by b().
This happened because the promising function is a separate task from the js code
in callPyObjectSuspending, so the sequence of events goes:

- enter main task,
    - enter callPyObjectSuspending(a)
        - enter promisingApply(a)
        - sets error flag and returns NULL
    - queue continue callPyObjectSuspending(a) in event loop
        now looks like [main task, continue callPyObjectSuspending(a)]

    - enter b()
        - enter Python
        - returns 7 with error state still set
    - rejects with "SystemError: <function b at 0x1140f20> returned a result with an exception set"
- queue continue main() in event loop
- continue callPyObjectSuspending(a)
    - pythonexc2js called attempting to read error flag set by promisingApply(a), fails with
        PythonError: TypeError: Pyodide internal error: no exception type or value

The solution: at the end of `_pyproxy_apply_promising` we move the error
flag into errStatus argument. In callPyObjectSuspending when we're ready we
move the error back from the errorStatus variable into the error flag before
calling `pythonexc2js()`