Rename callSyncifying to callPromising, add callOpts
Since we renamed `syncify` the method name `callSyncifying` doesn't make much
sense anymore. It is implemented in terms of a so-called `promising` wasm call,
so this lines up with that. OTOH:

* Nobody is likely to know what that refers to
* most people won't need this anyways
* it does return a promise so the name is somewhat descriptive
* the main point of it is to call a sync function and enable stack switching
* so maybe something like `suspendableCall` would be more appropriate?

We now have three boolean parameters for a Js-to-Python call:

1. kwargs
2. promising
3. relaxed

So we'd need 8 functions to cover all combinations of these. Currently we only
have 6 of these. Rather than adding the two remaining combinations which will
have annoying names, I added `callOptions` which takes an options argument as
the first argument. Despite the fact that options usually go as the last
argument, I think it makes sense to use the first argument for this so that all
remaining args are passed on to Python. There's also a suggestion of adding a
timeout parameter #4537 which this would reasoanbly accomodate.

Is `callOptions` a good name?