Snapshots: allow snapshotting some user code
This adds some basic ability to snapshot after executing user code. It is pretty
brittle right now:
1. It will crash if the user loads any binary extensions before taking the
snapshot
2. It doesn't track changes to the file system
3. It doesn't mind there being non-reproducible objects in the hiwire map

Snapshots will probably have to be experimental for quite a while.

1. I think I have a pretty good solution for this, which I can work on in a
followup.

2. One possibility here is we could serialize the entire filesystem state into
the memory snapshot. This would be hard and make the snapshot big, but we
wouldn't have to load python_stdlib.zip when restoring from a snapshot so it
probably wouldn't increase the total download size by much...

3. Should be pretty easy to fix.